RDoc-2566 Move "Delete by query" to be under "Common" operations folder - part 3 (move code files + fix links)
