+++ Documentation/5.2/Raven.Documentation.Pages/client-api/rest-api/queries/patch-by-query.markdown
@@ -131,7 +131,7 @@ double quotes within the request body using a backslash: `"` -> `\"`.
 ##### REST API  
 
 - [Query the Database](../../../client-api/rest-api/queries/query-the-database)  
-- [Delete by Query](../../../client-api/rest-api/queries/delete-by-query)  
+- [Delete by Query](../../../client-api/rest-api/queries/common/delete-by-query)  
 <br/>
 ### Indexes
