+++ Documentation/5.2/Samples/java/src/test/java/net/ravendb/ClientApi/Operations/Common/DeleteByQuery.java
